Remove `AddProductForm` component: Delete the unused dynamic product creation and editing form, including all associated logic for tabbed navigation, validation, SKU generation, pricing, variants, media, and stock management.
