+++ app/dashboard/products/[id]/edit/page.jsx
@@ -1,7 +1,8 @@
 import {BackBtn} from "@/components/ui/buttons";
 import {getCategoriesForCurrentOrganization} from "@/lib/queryCategories";
-import ProductForm from "@/components/dashboard/products/ProductForm";
 import {getProductById} from "@/lib/products/queryProducts";
+import {notFound} from "next/navigation";
+import ProductForm from "@/components/dashboard/products/ProductForm";
 
 export default async function Page({params}) {
     const {id} = await params;
@@ -9,9 +10,11 @@ export default async function Page({params}) {
     const product = await getProductById(id);
     const categories = await getCategoriesForCurrentOrganization();
 
+    if (!product) notFound();
+
     return (
         <div className={`min-h-screen`}>
-            <div className={`max-w-6xl mx-auto px-4 sm:px-6 lg:px-8 py-8`}>
+            <div className={`max-w-6xl mx-auto`}>
                 {/* Header */}
                 <div className={`flex items-center justify-between mb-8`}>
                     <div className={`flex items-center`}>
