+++ components/dashboard/products/ProductForm.jsx
@@ -1,28 +1,30 @@
 'use client'
 
-import {useRef, useState} from "react";
-import {CheckIcon, ExclamationTriangleIcon, InformationCircleIcon} from "@heroicons/react/24/outline";
-import BasicInfoForm from "@/components/dashboard/products/BasicInfoForm";
-import PricingForm from "@/components/dashboard/products/PricingForm";
-import InventoryForm from "@/components/dashboard/products/InventoryForm";
-import PhysicalForm from "@/components/dashboard/products/PhysicalForm";
-import MediaForm from "@/components/dashboard/products/MediaForm";
-import Link from "next/link";
-import {productCreationTips, productsFormTabs} from "@/data/constants/constants";
-import {toast} from "react-toastify";
-import {useRouter} from "next/navigation";
+import {useState} from 'react';
+import {CheckCircleIcon, CheckIcon, ExclamationTriangleIcon} from '@heroicons/react/24/outline';
+import {productFormTabs} from "@/utils/productUtils";
+import {showErrorToast, showSuccessToast} from "@/utils/toastUtil";
+import {productFormSchema} from "@/schema/productSchema";
 import {addProduct, updateProduct} from "@/lib/products/queryProducts";
+import {useRouter} from "next/navigation";
+import {ProgressLoader} from "@/components";
+import BasicInfoTab from "@/components/dashboard/products/BasicInfoTab";
+import PricingTab from "@/components/dashboard/products/PricingTab";
+import VariantsTab from "@/components/dashboard/products/VariantsTab";
+import MediaTab from "@/components/dashboard/products/MediaTab";
 
 export default function ProductForm({categories, product = null}) {
     const router = useRouter();
     const isEditing = !!product?.id;
 
+    const [loading, setLoading] = useState(false);
+    const [activeTab, setActiveTab] = useState('basic');
     const [formData, setFormData] = useState({
         sku: product?.sku || '',
         name: product?.name || '',
         description: product?.description || '',
-        category_id: product?.category_id || '',
         brand: product?.brand || '',
+        category_id: product?.category_id || '',
         unit_of_measure: product?.unit_of_measure || 'pcs',
         barcode: product?.barcode || '',
         qr_code: product?.qr_code || '',
@@ -33,113 +35,55 @@ export default function ProductForm({categories, product = null}) {
         reorder_point: product?.reorder_point || '',
         reorder_quantity: product?.reorder_quantity || '',
         weight: product?.weight || '',
-        dimensions: product?.dimensions || {
-            length: '',
-            width: '',
-            height: ''
-        },
+        dimensions: product?.dimensions || {length: '', width: '', height: ''},
         tags: product?.tags || [],
         is_active: product?.is_active || false,
-        is_trackable: product?.is_trackable || false,
+        is_trackable: product?.is_trackable || false
     });
 
-    const [images, setImages] = useState([]);
-    const [currentTag, setCurrentTag] = useState('');
-    const [errors, setErrors] = useState({});
-    const [isSubmitting, setIsSubmitting] = useState(false);
+    const [variants, setVariants] = useState(product?.variants || []);
+    const [images] = useState(product?.images || []);
+    const [touchedTabs, setTouchedTabs] = useState({basic: true});
 
-    const [activeTab, setActiveTab] = useState('basic');
-    const fileInputRef = useRef(null);
-
-    const handleInputChange = (field, value) => {
-        if (field.includes('.')) {
-            const [parent, child] = field.split('.');
-            setFormData(prev => ({
-                ...prev,
-                [parent]: {
-                    ...prev[parent],
-                    [child]: value
-                }
-            }));
-        } else {
-            setFormData(prev => ({...prev, [field]: value}));
+    // Tab completion status
+    const getTabStatus = (tabId) => {
+        if (tabId === 'basic') {
+            return formData.name && formData.sku;
         }
-
-        // Clear error when the user starts typing
-        if (errors[field]) {
-            setErrors(prev => ({...prev, [field]: null}));
+        if (tabId === 'pricing') {
+            return formData.cost_price && formData.selling_price;
         }
+        return true;
     };
 
-    const handleImageUpload = (e) => {
-        const files = Array.from(e.target.files);
-        files.forEach(file => {
-            const reader = new FileReader();
-            reader.onload = (e) => {
-                setImages(prev => [...prev, {
-                    id: Date.now() + Math.random(),
-                    file: file,
-                    preview: e.target.result,
-                    name: file.name
-                }]);
-            };
-            reader.readAsDataURL(file);
-        });
+    const changeTab = (tabId) => {
+        setActiveTab(tabId);
+        setTouchedTabs(prev => ({...prev, [tabId]: true}));
     };
 
-    const removeImage = (imageId) => {
-        setImages(prev => prev.filter(img => img.id !== imageId));
-    };
-
-    const addTag = () => {
-        if (currentTag.trim() && !formData.tags.includes(currentTag.trim())) {
-            setFormData(prev => ({
-                ...prev,
-                tags: [...prev.tags, currentTag.trim()]
-            }));
-            setCurrentTag('');
-        }
-    };
-
-    const removeTag = (tagToRemove) => {
-        setFormData(prev => ({
-            ...prev,
-            tags: prev.tags.filter(tag => tag !== tagToRemove)
-        }));
-    };
-
-    const validateForm = () => {
-        const newErrors = {};
-
-        if (!formData.sku.trim()) newErrors.sku = 'SKU is required';
-        if (!formData.name.trim()) newErrors.name = 'Product name is required';
-        if (!formData.category_id) newErrors.category_id = 'Category is required';
-        if (!formData.cost_price || parseFloat(formData.cost_price) < 0) {
-            newErrors.cost_price = 'Valid cost price is required';
-        }
-        if (!formData.selling_price || parseFloat(formData.selling_price) < 0) {
-            newErrors.selling_price = 'Valid selling price is required';
-        }
-        if (parseFloat(formData.selling_price) < parseFloat(formData.cost_price)) {
-            newErrors.selling_price = 'Selling price should be greater than cost price';
-        }
-
-        setErrors(newErrors);
-        return Object.keys(newErrors).length === 0;
+    const updateFormData = (field, value) => {
+        setFormData(prev => ({...prev, [field]: value}));
     };
 
     const handleSubmit = async (e) => {
         e.preventDefault();
 
-        if (!validateForm()) {
+        // Check required productFormTabs
+        const requiredTabs = productFormTabs.filter(tab => tab.required);
+        const incompleteTabs = requiredTabs.filter(tab => !getTabStatus(tab.id));
+
+        if (incompleteTabs.length > 0) {
+            showErrorToast(`Please complete the following required sections: ${incompleteTabs.map(t => t.name).join(', ')}`);
+            setActiveTab(incompleteTabs[0].id);
             return;
         }
 
-        setIsSubmitting(true);
+        setLoading(true);
 
         try {
             const submitData = {
                 ...formData,
+                variants,
                 cost_price: parseFloat(formData.cost_price),
                 selling_price: parseFloat(formData.selling_price),
                 min_stock_level: parseInt(formData.min_stock_level.trim() ? formData.min_stock_level.trim() : 0),
@@ -147,23 +91,27 @@ export default function ProductForm({categories, product = null}) {
                 reorder_point: parseInt(formData.reorder_point.trim() ? formData.reorder_point.trim() : 0),
                 reorder_quantity: parseInt(formData.reorder_quantity.trim() ? formData.reorder_quantity.trim() : 0),
                 weight: parseFloat(formData.weight.trim() ? formData.weight.trim() : 0),
-                images: images.map(img => img.preview)
+            };
+
+            const result = productFormSchema.safeParse(submitData);
+            if (!result.success) {
+                const errorMessages = JSON.parse(result.error.message);
+                for (let i = 0; i < errorMessages.length; i++) {
+                    showErrorToast(errorMessages[i].message);
+                }
+                return;
             }
 
             if (isEditing) {
                 await handleEditSubmit(submitData);
-            } else {
-                await handleCreateSubmit(submitData);
             }
+
+            await handleCreateSubmit(submitData);
         } catch (error) {
             console.error('Error adding product:', error);
-            toast.error(`An error has occurred!. Please try again`, {
-                    theme: 'dark',
-                    autoClose: 5000,
-                }
-            );
+            showErrorToast('An error has occurred!. Please try again');
         } finally {
-            setIsSubmitting(false);
+            setLoading(false);
         }
     };
 
@@ -171,14 +119,11 @@ export default function ProductForm({categories, product = null}) {
         const editedProduct = await updateProduct(product.id, submitData);
 
         if (!editedProduct) {
-            throw new Error('Failed to edit product');
+            showErrorToast('Failed to edit product');
+            return;
         }
 
-        toast.success(`Product ${editedProduct?.name} updated successfully`, {
-                theme: 'dark',
-                autoClose: 5000,
-            }
-        );
+        showSuccessToast(`Product ${editedProduct?.name} updated successfully`)
 
         router.push(`/dashboard/products`);
     }
@@ -187,184 +132,147 @@ export default function ProductForm({categories, product = null}) {
         const addedProduct = await addProduct(submitData);
 
         if (!addedProduct) {
-            throw new Error('Failed to add product');
+            showErrorToast('Failed to add product');
+            return;
         }
 
-        toast.success(`Product ${addedProduct?.name} created successfully`, {
-                theme: 'dark',
-                autoClose: 5000,
-            }
-        );
+        showSuccessToast(`Product ${addedProduct?.data.product?.name} created successfully`);
 
         router.push(`/dashboard/products`);
     }
 
-    const nextTab = () => {
-        const currentIndex = productsFormTabs.findIndex(tab => tab.id === activeTab);
-        const nextIndex = (currentIndex + 1) % productsFormTabs.length;
-        setActiveTab(productsFormTabs[nextIndex].id);
-    };
-
-    const prevTab = () => {
-        const currentIndex = productsFormTabs.findIndex(tab => tab.id === activeTab);
-        const prevIndex = (currentIndex - 1 + productsFormTabs.length) % productsFormTabs.length;
-        setActiveTab(productsFormTabs[prevIndex].id);
-    }
-
     return (
-        <form>
-            <div className={`flex flex-col lg:flex-row gap-8`}>
-                {/* Sidebar Navigation */}
-                <div className={`lg:w-64 flex-shrink-0`}>
-                    <div className={`bg-gray-800 rounded-xl p-4 border border-gray-700 sticky top-8`}>
-                        <nav className={`space-y-2`}>
-                            {productsFormTabs.map((tab) => {
-                                const Icon = tab.icon;
-                                return (
-                                    <button
-                                        key={tab.id}
-                                        type={`button`}
-                                        onClick={() => setActiveTab(tab.id)}
-                                        className={`w-full flex items-center px-4 py-3 text-left rounded-lg transition-colors ${
-                                            activeTab === tab.id
-                                                ? 'bg-teal-600 text-white'
-                                                : 'text-gray-400 hover:text-white hover:bg-gray-700'
-                                        }`}
-                                    >
-                                        <Icon className={`size-5 mr-3`}/>
-                                        {tab.name}
-                                    </button>
-                                );
-                            })}
-                        </nav>
+        <form onSubmit={handleSubmit}>
+            {/* Tabs */}
+            <div className={`border-b border-gray-800 mb-8 -mx-4 px-4 sm:mx-0 sm:px-0`}>
+                <nav className={`flex space-x-2 sm:space-x-8 overflow-x-auto scrollbar-hide`}>
+                    {productFormTabs.map((tab) => {
+                        const Icon = tab.icon;
+                        const isComplete = getTabStatus(tab.id);
+                        const isTouched = touchedTabs[tab.id];
 
-                        {/* Progress Indicator */}
-                        <div className={`mt-6 pt-6 border-t border-gray-700`}>
-                            <div className={`flex items-center text-sm text-gray-400`}>
-                                <InformationCircleIcon className={`size-4 mr-2`}/>
-                                <span>Fill required fields (*)</span>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-
-                {/* Main Content */}
-                <div className={`flex-1`}>
-                    <div className={`bg-gray-800 rounded-xl border border-gray-700 p-8`}>
-                        {activeTab === 'basic' && BasicInfoForm({
-                            formData,
-                            errors,
-                            onChange: handleInputChange,
-                            categories,
-                        })}
-                        {activeTab === 'pricing' && PricingForm({
-                            formData,
-                            errors,
-                            onChange: handleInputChange
-                        })}
-                        {activeTab === 'inventory' && InventoryForm({formData, onChange: handleInputChange})}
-                        {activeTab === 'physical' && PhysicalForm({formData, onChange: handleInputChange})}
-                        {activeTab === 'media' && MediaForm({
-                            formData,
-                            images,
-                            fileInputRef,
-                            handleImageUpload,
-                            removeImage,
-                            addTag,
-                            removeTag,
-                            currentTag,
-                            setCurrentTag
-                        })}
-                    </div>
-
-                    {/* Action Buttons */}
-                    <div className={`flex justify-between items-center mt-8`}>
-                        <Link
-                            href={`/dashboard/products`}
-                            className={`px-6 py-3 border border-gray-600 text-gray-400 hover:text-text hover:border-gray-500 rounded-lg transition-colors`}
-                        >
-                            Cancel
-                        </Link>
-
-                        <div className={`flex gap-4`}>
-                            {activeTab !== 'basic' && (
-                                <button
-                                    type={`button`}
-                                    disabled={isSubmitting}
-                                    onClick={prevTab}
-                                    className={`px-6 py-3 bg-gray-700 hover:bg-gray-600 text-white rounded-lg transition-colors`}
-                                >
-                                    Previous
-                                </button>
-                            )}
-
+                        return (
                             <button
+                                key={tab.id}
                                 type={`button`}
-                                disabled={isSubmitting}
-                                onClick={activeTab === 'media' ? handleSubmit : nextTab}
-                                className={`px-8 py-3 bg-teal-600 hover:bg-teal-700 disabled:bg-teal-600/50 disabled:cursor-not-allowed text-white rounded-lg font-medium transition-colors flex items-center`}
+                                onClick={() => changeTab(tab.id)}
+                                className={`flex items-center py-4 px-3 sm:px-1 border-b-2 font-medium text-sm transition-colors whitespace-nowrap relative ${
+                                    activeTab === tab.id
+                                        ? 'border-teal-500 text-teal-500'
+                                        : 'border-transparent text-gray-400 hover:text-gray-300 hover:border-gray-700'
+                                }`}
                             >
-                                {isSubmitting ? (
-                                    <>
-                                        <svg className="animate-spin -ml-1 mr-3 h-5 w-5 text-white"
-                                             xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
-                                            <circle className="opacity-25" cx="12" cy="12" r="10"
-                                                    stroke="currentColor" strokeWidth="4"></circle>
-                                            <path className="opacity-75" fill="currentColor"
-                                                  d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                                        </svg>
-                                        {isEditing ? 'Updating Product...' : 'Creating Product...'}
-                                    </>
-                                ) : (
-                                    activeTab === 'media' ? (
-                                        <>
-                                            <CheckIcon className={`size-5 mr-2`}/>
-                                            {isEditing ? 'Update Product' : 'Create Product'}
-                                        </>
-                                    ) : (
-                                        <span>Next</span>
-                                    )
+                                <Icon className={`size-5 mr-2 flex-shrink-0`}/>
+                                <span className={`hidden sm:inline`}>{tab.name}</span>
+                                <span className={`sm:hidden`}>{tab.name.split(' ')[0]}</span>
+                                {tab.required && (
+                                    <span className={`ml-1 text-red-400`}>*</span>
+                                )}
+                                {isTouched && isComplete && (
+                                    <CheckCircleIcon className={`size-4 ml-2 text-green-500`}/>
+                                )}
+                                {isTouched && !isComplete && tab.required && (
+                                    <ExclamationTriangleIcon className={`size-4 ml-2 text-yellow-500`}/>
                                 )}
                             </button>
-                        </div>
-                    </div>
+                        );
+                    })}
+                </nav>
+            </div>
 
-                    {/* Form Validation Summary */}
-                    {Object.keys(errors).length > 0 && (
-                        <div className={`mt-6 p-4 bg-red-900/20 border border-red-500/50 rounded-lg`}>
-                            <div className={`flex items-center mb-2`}>
-                                <ExclamationTriangleIcon className={`size-5 text-red-400 mr-2`}/>
-                                <h3 className={`text-red-400 font-medium`}>Please fix the following errors:</h3>
-                            </div>
-                            <ul className={`text-red-300 text-sm space-y-1 ml-7`}>
-                                {Object.entries(errors).map(([field, error]) => (
-                                    <li key={field}>• {error}</li>
-                                ))}
-                            </ul>
-                        </div>
-                    )}
+            {activeTab === 'basic' && (
+                <BasicInfoTab
+                    formData={formData}
+                    setFormData={setFormData}
+                    updateFormData={updateFormData}
+                    categories={categories}
+                />
+            )}
 
-                    <HelpSection/>
+            {/* Pricing & Stock Tab */}
+            {activeTab === 'pricing' && (
+                <PricingTab formData={formData} updateProduct={updateFormData}/>
+            )}
+
+            {/* Variants Tab */}
+            {activeTab === 'variants' && (
+                <VariantsTab variants={variants} setVariants={setVariants}/>
+            )}
+
+            {/* Media Tab */}
+            {activeTab === 'media' && (
+                <MediaTab images={images}/>
+            )}
+
+            {/* Navigation & Action Buttons */}
+            <div className={`flex flex-wrap gap-4 justify-between items-center mt-8`}>
+                <button
+                    type={`button`}
+                    onClick={() => {
+                        const currentIndex = productFormTabs.findIndex(t => t.id === activeTab);
+                        if (currentIndex > 0) {
+                            changeTab(productFormTabs[currentIndex - 1].id);
+                        }
+                    }}
+                    disabled={activeTab === 'basic'}
+                    className={`px-6 py-3 bg-gray-800 hover:bg-gray-700 rounded-lg font-medium transition-colors border border-gray-700 disabled:opacity-50 disabled:cursor-not-allowed`}
+                >
+                    Previous
+                </button>
+
+                {activeTab !== 'media' && (
+                    <button
+                        type={`button`}
+                        onClick={() => {
+                            const currentIndex = productFormTabs.findIndex(t => t.id === activeTab);
+                            if (currentIndex < productFormTabs.length - 1) {
+                                changeTab(productFormTabs[currentIndex + 1].id);
+                            }
+                        }}
+                        className={`px-6 py-3 bg-teal-600 hover:bg-teal-700 rounded-lg font-medium transition-colors`}
+                    >
+                        Next
+                    </button>
+                )}
+
+                {activeTab === 'media' && (
+                    <button
+                        type={`submit`}
+                        disabled={loading}
+                        className={`px-8 py-3 bg-teal-600 hover:bg-teal-700 disabled:bg-teal-600/50 disabled:cursor-not-allowed rounded-lg font-medium transition-colors flex items-center`}
+                    >
+                        {loading ? (
+                            <>
+                                <ProgressLoader className={`mr-2`}/>
+                                {isEditing ? 'Updating Product...' : 'Creating Product...'}
+                            </>
+                        ) : (
+                            <>
+                                <CheckIcon className={`size-5 mr-2`}/>
+                                {isEditing ? 'Update Product' : 'Create Product'}
+                            </>
+                        )}
+                    </button>
+                )}
+            </div>
+
+            {/* Progress indicator */}
+            <div className={`mt-6 bg-gray-900 rounded-lg p-4 border border-gray-800`}>
+                <div className={`flex justify-between items-center mb-2`}>
+                    <span className={`text-sm font-medium text-gray-300`}>Completion Progress</span>
+                    <span className={`text-sm text-gray-400`}>
+                        {productFormTabs.filter(tab => getTabStatus(tab.id)).length} / {productFormTabs.filter(tab => tab.required).length} required sections
+                    </span>
+                </div>
+                <div className={`w-full bg-gray-800 rounded-full h-2`}>
+                    <div
+                        className={`bg-teal-600 h-2 rounded-full transition-all duration-300`}
+                        style={{
+                            width: `${(productFormTabs.filter(tab => tab.required && getTabStatus(tab.id)).length / productFormTabs.filter(tab => tab.required).length) * 100}%`
+                        }}
+                    ></div>
                 </div>
             </div>
         </form>
-    )
-}
-
-function HelpSection() {
-    return (
-        <div className={`mt-8 p-6 bg-blue-900/20 border border-blue-500/50 rounded-lg`}>
-            <div className={`flex items-start`}>
-                <InformationCircleIcon className={`size-6 text-blue-400 mt-0.5 mr-3 flex-shrink-0`}/>
-                <div>
-                    <h3 className={`text-blue-400 font-medium mb-2`}>Product Creation Tips</h3>
-                    <ul className={`text-blue-300 text-sm space-y-1 list-disc list-inside`}>
-                        {productCreationTips.map((tip, index) => (
-                            <li key={index}>{tip}</li>
-                        ))}
-                    </ul>
-                </div>
-            </div>
-        </div>
-    )
+    );
 }
