+++ lib/products/queryProducts.js
@@ -28,12 +28,9 @@ export async function getProductById(productId) {
         .limit(1)
         .single();
 
-    if (error) {
-        console.error('Error fetching product:', error);
-        return null;
-    }
+    if (error) throw error;
 
-    return product;
+    return product || null;
 }
 
 export async function getCategoriesForProducts() {
@@ -106,12 +103,10 @@ export async function addProduct(data) {
 
 export async function updateProduct(productId, data) {
     const supabase = await createClient();
-    const organizationId = await getCurrentOrganizationId();
 
     const {data: product, error} = await supabase
         .from('products')
         .update({
-            organization_id: organizationId,
             sku: data.sku,
             name: data.name,
             description: data.description,
@@ -138,12 +133,20 @@ export async function updateProduct(productId, data) {
         .select()
         .single();
 
-    if (error) {
-        console.error('Error updating product: ', error);
-        return null;
-    }
+    if (error) throw error;
 
-    return product;
+    const productVariants = data.variants.map(variant => ({
+        product_id: productId,
+        sku: variant.sku,
+        name: variant.name,
+        attributes: variant.attributes,
+        cost_price: variant.cost_price,
+        selling_price: variant.selling_price,
+        barcode: variant.barcode,
+        is_active: variant.is_active ?? true
+    }));
+
+    return product || null;
 }
 
 export async function deleteProduct(productId) {
